refactor(functions): extract deployFunction poll loop into functions_deploy
